+++ src/platform/android.rs
@@ -1,31 +1,30 @@
 #![cfg(any(target_os = "android"))]
 
-use crate::{EventLoop, Window, WindowBuilder};
-use std::os::raw::c_void;
+use crate::event_loop::{EventLoop, EventLoopWindowTarget};
+use crate::window::{Window, WindowBuilder};
 
 /// Additional methods on `EventLoop` that are specific to Android.
-pub trait EventLoopExtAndroid {
-    /// Makes it possible for glutin to register a callback when a suspend event happens on Android
+pub trait EventLoopExtAndroid {}
+
+impl<T> EventLoopExtAndroid for EventLoop<T> {}
+
+/// Additional methods on `EventLoopWindowTarget` that are specific to Android.
+pub trait EventLoopWindowTargetExtAndroid {
+    /// Makes it possible for to register a callback when a suspend event
+    /// happens on Android.
     fn set_suspend_callback(&self, cb: Option<Box<dyn Fn(bool) -> ()>>);
 }
 
-impl EventLoopExtAndroid for EventLoop {
+impl<T> EventLoopWindowTargetExtAndroid for EventLoopWindowTarget<T> {
     fn set_suspend_callback(&self, cb: Option<Box<dyn Fn(bool) -> ()>>) {
-        self.event_loop.set_suspend_callback(cb);
+        self.p.set_suspend_callback(cb);
     }
 }
 
 /// Additional methods on `Window` that are specific to Android.
-pub trait WindowExtAndroid {
-    fn native_window(&self) -> *const c_void;
-}
+pub trait WindowExtAndroid {}
 
-impl WindowExtAndroid for Window {
-    #[inline]
-    fn native_window(&self) -> *const c_void {
-        self.window.native_window()
-    }
-}
+impl WindowExtAndroid for Window {}
 
 /// Additional methods on `WindowBuilder` that are specific to Android.
 pub trait WindowBuilderExtAndroid {}